pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "myst_spec_py"
authors = [{name = "Chris Sewell", email = "chrisj_sewell@hotmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = ["markdown-it-py~=2.0.1"]

[project.urls]
Home = "https://github.com/chrisjsewell/commonmark-ast-spec"

[project.scripts]
cmark-to-mdast = "myst_spec_py.__main__:cli_cmark_to_mdast"

[project.optional-dependencies]
testing = ["pytest"]

[tool.isort]
profile = "black"
force_sort_within_sections = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py39

[testenv]
usedevelop=True

[testenv:py{38,39,310}]
description = Run the test suite
extras = testing
deps = black
commands = pytest {posargs:tests/}

[testenv:cli]
description = Run the CLI
commands =
    python -m myst_spec_py {posargs}
"""
